Uncaught TypeError : Fixed #41
E_ERROR (1): Uncaught TypeError: Unsupported operand types: int + string in /wp-content/plugins/woocommerce-product-stock-alert/classes/class-woo-product-stock-alert-ajax.php:220 
This issue fixed.

To avoid this error please cast the variable you get the number of subscribers via get_post_meta as integer. Sometime the DB only returns an empty string instead of an integer.

To avoid this error please cast the variable you get the number of subscribers via get_post_meta as integer. Sometime the DB only returns an empty string instead of an integer.
To avoid this error we cast the variable that get the number of subscribers via get-post-meta as integer. Sometime the DB only returns an empty string instead of an integer.
So adding
$no_of_subscribers = intval($no_of_subscribers);